requirements.txt
psycopg2-binary>=2.9.9
bcrypt>=4.0.1